Basic functionality achieved with vsss_rs library. Some notes:
We don't support any way to serialize an arbitrary string into a field element, suitable for secret sharing
We don't check the salted hash of the secret key.
Upon initial upload, the secret key is split into a random polynomial. The password and password guess shamir polynomials are deterministically generated, by seeding the rng with a constant.